little bit of re-organizing in the hope that it makes things easier for others
